docker fixes

+++ data-analysis/run_app.sh
@@ -1,8 +1,15 @@
 #!/bin/bash
 
-# remove old html pages
-rm /web-app/event*.html
+if [ "$DATA_ANALYSIS" = "true" ]; then 
+    # remove old html pages
+    rm /web-app/event*.html
+    rm /web-app/results/sentiment/*.json
+    rm /web-app/results/world/*.json
 
-# python data_retrieval.py
-# python clean_and_sentiment.py
-python prepare_app.py
+    # python data_retrieval.py
+    # python clean_and_sentiment.py
+    # python topic_analysis.py
+    python prepare_app.py
+else 
+    echo "Bypassing data analysis";
+fi
